yolo-api/requirements.txt
fastapi
uvicorn
onnxruntime
pillow
torchvision